src/initialization/requirements.txt
pyyaml
pandas
openai
shapely
openpyxl
astunparse
scipy